refactor: rename file System.hpp to FileInterface.hpp
